+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Next
+- add property `backgroundColor` in `TalkDialog`.
+
 # 1.2.0
 - add `MouseGesture` mixin to listen mouse gestures see [documentation](https://bonfire-engine.github.io/#/gestures?id=mousegesture)
 - add method `worldPositionToScreen` in `BonfireGame`.