+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # NEXT
 - improvements in `TiledWorldMap`
 - add property `backgroundColor` in `TalkDialog`.
+- performance improvements in `TiledWorldMap`
 
 # 1.2.0
 - add `MouseGesture` mixin to listen mouse gestures see [documentation](https://bonfire-engine.github.io/#/gestures?id=mousegesture)